refactor: separate json tests from brc-20 tests
